requirements.txt
# --- ML Core ---
numpy==2.0.2
scipy==1.13.1
scikit-learn==1.6.1
spacy==3.8.11
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl

# --- Server & API ---
fastapi==0.128.0
uvicorn[standard]==0.39.0
pydantic[email]==2.12.5
pydantic-settings==2.11.0
python-multipart==0.0.20
python-dotenv==1.2.1
httpx==0.28.1

# --- Utilities ---
boto3==1.42.35
PyPDF2==3.0.1
python-docx==1.2.0
pillow==11.3.0
lxml==6.0.2